requirements.txt
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
pandas==1.5.3
opencv-python==4.8.1.78
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
Pillow==10.0.1
Flask==2.3.3
Werkzeug==2.3.7
plotly==5.17.0
streamlit==1.28.1
gradio==3.50.2
albumentations==1.3.1
imgaug==0.4.0